importing image with fs is completed

+++ client/src/pages/dashboard/pages/Doctors.jsx
@@ -1,10 +1,13 @@
 import { useEffect, useState } from "react";
 import Table from "../../ui/Table";
+import Input from "../../ui/Input";
 import Layout from "../layout/Main";
 import Button from "../../ui/Button";
-import { Fetch } from "../../../constant/Fetch";
+import { Fetch } from "../../../utils/Fetch";
 import { NavLink } from "react-router-dom";
-import { Delete } from "../../../constant/Delete";
+import { Delete } from "../../../utils/Delete";
+import Modal from "../../ui/Modal";
+import { handleModelInput } from "../../../utils/handleInput";
 
 const PORT = import.meta.env.VITE_SERVER_API;
 const API = `${PORT}/getdoctors`;
@@ -13,25 +16,42 @@ const Doctor = () => {
   const { data, isLoading, error, getData } = Fetch();
   const { deleteData, setError, setIsLoading } = Delete();
 
-  const [isUpdateModalOpen, setUpdateModalOpen] = useState(false);
+  const [isModalOpen, setModalOpen] = useState(false);
+  const [doctorToDelete, setDoctorToDelete] = useState(null);
+  const [doctorNameInput, setDoctorNameInput] = useState("");
+  const [nameError, setNameError] = useState(null);
+
+  const toggleModal = () => {
+    setModalOpen((prev) => !prev);
+    setDoctorNameInput("");
+    setNameError(null);
+  };
 
   useEffect(() => {
     getData(API);
   }, []);
 
-  const toggleModal = () => {
-    setUpdateModalOpen(!isUpdateModalOpen);
+  const confirmDelete = (doctor) => {
+    setDoctorToDelete(doctor);
+    toggleModal();
   };
 
-  //  deleting data
-  const handleDelete = async (doctorID) => {
-    console.log(doctorID);
-    try {
-      const apiUrl = `${PORT}/deletedoctor/${doctorID}`;
-      await deleteData(apiUrl);
-      getData(API);
-    } catch (error) {
-      setError(error);
+  // Deleting doctor after name validation
+  const handleDelete = async () => {
+    if (doctorToDelete.Doctor_name === doctorNameInput) {
+      setIsLoading(true);
+      try {
+        const apiUrl = `${PORT}/deletedoctor/${doctorToDelete.id}`;
+        await deleteData(apiUrl);
+        getData(API);
+        toggleModal();
+      } catch (error) {
+        setError(error);
+      }
+    } else {
+      setNameError(
+        "Doctor's name doesn't match. Please enter the correct name."
+      );
     }
   };
 
@@ -55,19 +75,19 @@ const Doctor = () => {
         status: doctor.status,
         actions: (
           <div className="flex items-center gap-1">
-            <Button
-              className="py-2 px-3 rounded-full bg-primary text-white"
-              onClick={() => toggleModal(doctor)}
-            >
+            <Button className="py-2 px-3 rounded-full bg-primary text-white">
               <i className="fa-solid fa-pen"></i>
             </Button>
             <Button
               className="py-2 px-3 rounded-full bg-red-600 text-white"
-              onClick={() => handleDelete(doctor.id)}
+              onClick={() => confirmDelete(doctor)}
             >
               <i className="fa-solid fa-trash"></i>
             </Button>
-            <Button className="py-2 px-3 rounded-full bg-slate-400 text-white">
+            <Button
+              className="py-2 px-3 rounded-full bg-slate-400 text-white"
+              onClick={toggleModal}
+            >
               <i className="fa-solid fa-eye"></i>
             </Button>
           </div>
@@ -89,6 +109,48 @@ const Doctor = () => {
         {error && <p>Error: {error}</p>}
 
         {data && <Table columns={columns} data={tableData} />}
+
+        {isModalOpen && doctorToDelete && (
+          <Modal
+            onClose={toggleModal}
+            title="Delete Confirmation"
+            data={doctorToDelete}
+            footer={
+              <>
+                <Button
+                  className="bg-red-600 text-white py-2 px-4 rounded-full"
+                  onClick={handleDelete}
+                >
+                  Delete
+                </Button>
+                <Button
+                  className="bg-gray-600 text-white py-2 px-4 rounded-full"
+                  onClick={toggleModal}
+                >
+                  Cancel
+                </Button>
+              </>
+            }
+          >
+            <div className="flex flex-col gap-3">
+              <p>
+                Are you sure you want to delete
+                <span className="text-primary">
+                  {doctorToDelete.Doctor_name}?
+                </span>
+              </p>
+              <p>Write Down Doctor's Name to Confirm:</p>
+              <Input
+                type="text"
+                name="doctorNameInput"
+                placeHolder="Dr. Name Here"
+                value={doctorNameInput}
+                onChange={handleModelInput(setDoctorNameInput)}
+              />
+              {nameError && <p className="text-red-600">{nameError}</p>}
+            </div>
+          </Modal>
+        )}
       </div>
     </Layout>
   );
